sb4/requirements.txt
py-cord[voice]==2.6.1
anthropic>=0.40.0
python-dotenv>=1.0.0
SpeechRecognition>=3.10.0
PyNaCl>=1.5.0
